EZP-26179: Refactored session rest actions to their own controller
delete, create and refresh session actions are deprecated in the User controller,
and moved to a new one, SessionController.
